ObjectPool 제네릭으로 변경 ( Note에서만 쓰다가 추후 마디선 등 더 필요하기 때문에 중복 코드 방지 )

+++ Assets/Scripts/Globals/FileReader.cs
@@ -91,7 +91,7 @@ public class FileReader
                 {
                     string[] arr = line.Split( ',' );
                     string[] LNTiming = arr[5].Split( ':' );
-                    data.notes.Add( new MetaData.Notes( int.Parse( arr[0] ), int.Parse( arr[1] ), ( uint )int.Parse( arr[2] ), int.Parse( arr[3] ), int.Parse( LNTiming[0] ) ) );
+                    data.notes.Add( new MetaData.Notes( int.Parse( arr[0] ), int.Parse( arr[1] ), float.Parse( arr[2] ), int.Parse( arr[3] ), int.Parse( LNTiming[0] ) ) );
                 }
             }
         }
